+++ src/openkit-shared/Protocol/HTTPClient.cs
@@ -69,6 +69,7 @@ namespace Dynatrace.OpenKit.Protocol
         internal const string QueryKeyVersion = "va";
         internal const string QueryKeyPlatformType = "pt";
         internal const string QueryKeyAgentTechnologyType = "tt";
+        internal const string QueryKeyNewSession = "ns";
 
         // additional reserved characters for URL encoding
         private static readonly char[] QueryReservedCharacters = { '_' };
@@ -79,6 +80,7 @@ namespace Dynatrace.OpenKit.Protocol
 
         // URLs for requests
         private readonly string monitorURL;
+        private readonly string newSessionUrl;
         private readonly string timeSyncURL;
 
         private readonly int serverID;
@@ -91,6 +93,7 @@ namespace Dynatrace.OpenKit.Protocol
             this.logger = logger;
             serverID = configuration.ServerID;
             monitorURL = BuildMonitorURL(configuration.BaseURL, configuration.ApplicationID, configuration.ServerID);
+            newSessionUrl = BuildNewSessionUrl(configuration.BaseURL, configuration.ApplicationID, configuration.ServerID);
             timeSyncURL = BuildTimeSyncURL(configuration.BaseURL);
         }
 
@@ -114,7 +117,7 @@ namespace Dynatrace.OpenKit.Protocol
 
         public StatusResponse SendNewSessionRequest()
         {
-            return (StatusResponse)SendRequest(RequestType.NewSession, monitorURL, null, null, "GET") ?? new StatusResponse(logger, string.Empty, int.MaxValue, new Dictionary<string, List<string>>());
+            return (StatusResponse)SendRequest(RequestType.NewSession, newSessionUrl, null, null, "GET") ?? new StatusResponse(logger, string.Empty, int.MaxValue, new Dictionary<string, List<string>>());
         }
 
         // generic request send with some verbose output and exception handling
@@ -283,6 +286,15 @@ namespace Dynatrace.OpenKit.Protocol
             AppendQueryParam(monitorURLBuilder, QueryKeyAgentTechnologyType, ProtocolConstants.AgentTechnologyType);
 
             return monitorURLBuilder.ToString();
+        }
+
+        private string BuildNewSessionUrl(string baseUrl, string applicationId, int serverId)
+        {
+            var newSessionUrlBuilder = new StringBuilder(BuildMonitorURL(baseUrl, applicationId, serverId));
+
+            AppendQueryParam(newSessionUrlBuilder, QueryKeyNewSession, "1");
+
+            return newSessionUrlBuilder.ToString();
         }
 
         // build URL used for time sync requests