+++ tests/openkit-shared-tests/Protocol/HTTPClientTest.cs
@@ -43,6 +43,10 @@ namespace Dynatrace.OpenKit.Protocol
             + "&" + HTTPClient.QueryKeyVersion + "=" + ProtocolConstants.OpenKitVersion
             + "&" + HTTPClient.QueryKeyPlatformType + "=" + ProtocolConstants.PlatformTypeOpenKit
             + "&" + HTTPClient.QueryKeyAgentTechnologyType + "=" + ProtocolConstants.AgentTechnologyType;
+
+        private static readonly string NewSessionUrl = MonitorURL
+            + $"&{HTTPClient.QueryKeyNewSession}=1";
+
         private const string TimeSyncURL = BaseURL + "?" + HTTPClient.RequestTypeTimeSync;
 
         private static readonly HTTPClient.HTTPResponse StatusResponse = new HTTPClient.HTTPResponse
@@ -241,7 +245,7 @@ namespace Dynatrace.OpenKit.Protocol
         }
 
         [Test]
-        public void SendNewSessionRequestSendRequestToMonitorURL()
+        public void SendNewSessionRequestSendRequestToNewSessionURL()
         {
             // given
             HTTPClient target = spyClient;
@@ -256,7 +260,7 @@ namespace Dynatrace.OpenKit.Protocol
             Assert.That(obtained, Is.Not.Null);
             Assert.That(obtained.ResponseCode, Is.EqualTo(200));
 
-            spyClient.Received(1).DoGetRequest(MonitorURL, null);
+            spyClient.Received(1).DoGetRequest(NewSessionUrl, null);
         }
 
         [Test]
@@ -677,7 +681,7 @@ namespace Dynatrace.OpenKit.Protocol
         public void SendBeaconRequestReturnsAnUnknownErrorResponseForWrongHttpResponse()
         {
             // given
-            HTTPClient target = spyClient; 
+            HTTPClient target = spyClient;
             spyClient.WhenForAnyArgs(x => x.DoPostRequest(string.Empty, string.Empty, null)).DoNotCallBase();
             spyClient.DoPostRequest(string.Empty, string.Empty, null).ReturnsForAnyArgs(TimeSyncResponse);